backend/requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.9
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
psycopg[binary]>=3.0.0
alembic>=1.13.0

# AI
openai>=1.0.0
replicate>=0.25.0

# Rate limiting
slowapi>=0.1.9

# OAuth & Social (Этап 4)
cryptography>=42.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth>=2.0.0
httpx>=0.27.0
